rename: fcm 도메인을 notification으로 변경
포맷팅도 진행했습니다
