+++ src/main/java/com/umc/product/notification/adapter/in/web/dto/request/FcmTestSendRequest.java
@@ -1,6 +1,6 @@
-package com.umc.product.fcm.adapter.in.web.dto.request;
+package com.umc.product.notification.adapter.in.web.dto.request;
 
-import com.umc.product.fcm.application.port.in.NotificationCommand;
+import com.umc.product.notification.application.port.in.NotificationCommand;
 
 public record FcmTestSendRequest(Long memberId, String title, String body) {
