+++ src/main/java/com/umc/product/notification/application/port/in/NotificationCommand.java
@@ -1,4 +1,4 @@
-package com.umc.product.fcm.application.port.in;
+package com.umc.product.notification.application.port.in;
 
 
 public record NotificationCommand(Long memberId, String title, String body) {
